+++ src/com/melody/core/Game.java
@@ -13,6 +13,8 @@ public final class Game extends Canvas implements Runnable {
 	private Input _input;
 	private StatDisplay _stat = new StatDisplay();
 	
+	public boolean autoClearBg = true;
+	
 	public int backgroundColor = 0xFFFFFF;
 	
 	private long _deltaTime;
@@ -25,8 +27,10 @@ public final class Game extends Canvas implements Runnable {
 	
 	protected void paint(Graphics g) {
 		// clear canvas
-		g.setColor(backgroundColor);
-		g.fillRect(0, 0, getWidth(), getHeight());
+		if (autoClearBg) {
+			g.setColor(backgroundColor);
+			g.fillRect(0, 0, getWidth(), getHeight());
+		}
 		
 		// call render manual from scene
 //		// render child
@@ -62,6 +66,7 @@ public final class Game extends Canvas implements Runnable {
 			_currentScene = _newScene;
 			_currentScene.preInit();
 			_newScene = null;
+			autoClearBg = true;
 			
 			System.gc();
 		}
@@ -143,6 +148,10 @@ public final class Game extends Canvas implements Runnable {
 	
 	public final Input get_input() {
 		return _input;
+	}
+	
+	public final Thread get_thread() {
+		return _mainThread;
 	}
 
 }