+++ src/com/melody/core/MainEngine.java
@@ -51,7 +51,13 @@ public abstract class MainEngine extends MIDlet {
 		_gameRoot.setFullScreenMode(true);
 		_gameRoot.start();
 		
-		Display.getDisplay(this).setCurrent(_gameRoot);
+		Display display = Display.getDisplay(this);
+		display.setCurrent(_gameRoot);
+		
+		// Print device capability
+		if (enableStat) {
+			System.out.println(display.numAlphaLevels());
+		}
 		
 		handleGameReady();
 	}