+++ src/com/mlbb/scene/MainMenu.java
@@ -1,0 +1,28 @@
+package com.mlbb.scene;
+
+import javax.microedition.lcdui.Graphics;
+
+import com.melody.core.Scene;
+
+public class MainMenu extends Scene {
+
+	public MainMenu() {
+	}
+
+	public void initialize() {
+
+	}
+
+	public void update(long dt) {
+
+	}
+
+	public void render(Graphics g) {
+
+	}
+
+	public void destroy() {
+
+	}
+
+}