+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2021-12-26
+### Changed
+- BMFont now can calculate width
+- Simplify BMFont render process by one line
+- Auto clear background now can be toggled
+
+### Added
+- BMFont parser using nodejs (i love js)
+- mlbb intro section done
+
+
 ## [1.1.0] - 2021-12-23
 ### Changed
 - Fix Antenna build problem